requirements_conda.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
audioread==3.0.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.3
chatterbox-tts==0.1.4
click==8.2.1
conformer==0.3.2
Cython==3.1.4
decorator==5.2.1
Deprecated==1.2.18
diffusers==0.29.0
distlib==0.4.0
einops==0.8.1
fastapi==0.116.2
ffmpy==0.6.1
filelock==3.19.1
Flask==3.1.2
fsspec==2025.9.0
gradio==5.44.1
gradio_client==1.12.1
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
identify==2.6.14
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
jaconv==0.4.0
Jinja2==3.1.6
joblib==1.5.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
nodeenv==1.9.1
numba==0.61.2
numpy==1.25.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnx==1.19.0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pkuseg==0.0.25
platformdirs==4.4.0
pooch==1.8.2
pre_commit==4.3.0
protobuf==6.32.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pykakasi==2.3.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
resemble-perth==1.0.1
rich==14.1.0
ruff==0.13.0
s3tokenizer==0.2.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.2
scipy==1.16.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
starlette==0.48.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.20.3
tomlkit==0.13.3
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
transformers==4.46.3
triton==3.2.0
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.34.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
zipp==3.23.0
